Projekat II/analytics/requirements.txt
astroid==2.11.5
click==8.1.3
colorama==0.4.4
dill==0.3.5.1
Flask==2.1.2
grpcio==1.46.3
grpcio-tools==1.46.3
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
paho-mqtt==1.6.1
platformdirs==2.5.2
protobuf==3.20.1
pylint==2.14.0
six==1.16.0
tomli==2.0.1
tomlkit==0.11.0
Werkzeug==2.1.2
wrapt==1.14.1
